delete forms

+++ resources/views/components/forms/delete-category.blade.php
@@ -5,7 +5,7 @@
                     {{ __('Cancel') }}
                 </x-secondary-button>
 
-                <x-primary-button @click="handleSubmit">
+                <x-primary-button @click="handleSubmit" class="justify-center">
                     <span x-show="sending">
                         <x-spinner />
                     </span>
@@ -25,29 +25,30 @@
             async handleSubmit(event) {
 
                 this.$event.preventDefault()
-                this.sending = true
+                if (!this.sending) {
+                    this.sending = true
 
-                await fetch('/api/categories/' + this.$store.api.queryId, {
-                    method: 'DELETE',
-                    headers: {
-                        'Content-Type': 'application/json',
-                        'X-CSRF-TOKEN': this.$store.csrf.token,
-                    },
-                })
-                .then((response) => response.json())
-                .then(data => {
-                    console.log(data)
-                    if (data.errors) {
-                        this.error = data.message
-                    } else {
-                        this.sending = false
-                        window.history.back();
-                    }
-                })
-                .catch((error) => {
-                    console.log(error)
-                });
-
+                    await fetch('/api/categories/' + this.$store.api.queryId, {
+                        method: 'DELETE',
+                        headers: {
+                            'Content-Type': 'application/json',
+                            'X-CSRF-TOKEN': this.$store.csrf.token,
+                        },
+                    })
+                    .then((response) => response.json())
+                    .then(data => {
+                        console.log(data)
+                        if (data.errors) {
+                            this.error = data.message
+                        } else {
+                            this.sending = false
+                            window.history.back();
+                        }
+                    })
+                    .catch((error) => {
+                        console.log(error)
+                    });
+                }
             },
 
         }))
